refactor: migrate email capture state to Zustand and local useState
- Replaced React Context with Zustand for shared showEmailBar state.
- Moved form-specific state (email, status) to local useState in EmailForm.
- Removed EmailCaptureContext and useEmailCaptureLogic hook.
- Consolidated STORAGE_KEY into the email store.
- Updated App.tsx, MainLayout.tsx, and related components to use the new state management.
