+++ src/App.tsx
@@ -3,15 +3,14 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { lazy, Suspense } from 'react';
+import { lazy, Suspense, useEffect } from 'react';
 import { Outlet, useLocation } from 'react-router-dom';
 import { AnimatePresence, motion } from 'motion/react';
 import { MainLayout } from './layouts/MainLayout';
 import { motionTokens } from './styles/motion';
 import { PageSkeleton } from './components/ui/PageSkeleton';
-import { EmailCaptureProvider } from './features/email-capture/EmailCaptureContext';
 import { NewsletterBanner } from './features/email-capture/NewsletterBanner';
-import { useEmailCaptureLogic } from './hooks/useEmailCaptureLogic';
+import { useEmailStore, STORAGE_KEY } from './features/email-capture/emailStore';
 
 import { Box } from './layouts/Primitives';
 
@@ -26,12 +25,26 @@ const BlogPost = lazy(() => import('./pages/BlogPost'));
 const About = lazy(() => import('./pages/About'));
 const Contact = lazy(() => import('./pages/Contact'));
 
+const BANNER_DELAY_MS = 30000; // 30s delay
+
 export function RootLayout() {
   const location = useLocation();
-  const emailLogic = useEmailCaptureLogic();
+  const showEmailBar = useEmailStore((state) => state.showEmailBar);
+  const setShowEmailBar = useEmailStore((state) => state.setShowEmailBar);
+
+  useEffect(() => {
+    const isDismissed = sessionStorage.getItem(STORAGE_KEY) === 'true';
+    if (isDismissed) return;
+
+    const timer = setTimeout(() => {
+      setShowEmailBar(true);
+    }, BANNER_DELAY_MS);
+
+    return () => clearTimeout(timer);
+  }, [setShowEmailBar]);
 
   return (
-    <EmailCaptureProvider {...emailLogic}>
+    <>
       <MainLayout>
         <AnimatePresence mode="wait">
           <Box
@@ -50,9 +63,9 @@ export function RootLayout() {
         </AnimatePresence>
       </MainLayout>
       <AnimatePresence>
-        {emailLogic.showEmailBar && <NewsletterBanner />}
+        {showEmailBar && <NewsletterBanner />}
       </AnimatePresence>
-    </EmailCaptureProvider>
+    </>
   );
 }
