+++ src/features/email-capture/NewsletterBanner.tsx
@@ -3,11 +3,11 @@ import { EmailForm } from './EmailForm';
 import { Mail, X } from 'lucide-react';
 import { motionTokens } from '@/styles/motion';
 import { motion } from 'motion/react';
-import { useEmailCaptureContext } from './EmailCaptureContext';
+import { useEmailStore } from './emailStore';
 import { Button } from '@/layouts/Primitives';
 
 export function NewsletterBanner() {
-  const { hideBar } = useEmailCaptureContext();
+  const hideBar = useEmailStore((state) => state.hideBar);
 
   return (
     <Box
